requirements.txt
sounddevice>=0.4.6
webrtcvad>=2.0.10
pyttsx3>=2.90
faster-whisper>=0.9.0
gTTS>=2.3.2
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
flask>=3.0.0 